削除機能　修正２

+++ app/controllers/items_controller.rb
@@ -35,11 +35,12 @@ class ItemsController < ApplicationController
    end
 
    def destroy
-    current_user.id == @item.user_id
+    if current_user.id == @item.user_id
      if @item.destroy
       redirect_to root_path
      end
    end
+  end
   private
   def item_params
   params.require(:item).permit(:image,:name, :category_id, :status_id, :shippingcharges_id , :sender_id, :scheduleddelivery_id, :price, :description).merge(user_id: current_user.id)
